# Globber rewrite