--- a/src/Cake.Core.Tests/Unit/IO/GlobberTests.cs
+++ b/src/Cake.Core.Tests/Unit/IO/GlobberTests.cs
@@ -38,180 +38,234 @@ namespace Cake.Core.Tests.Unit.IO
 
         public sealed class TheMatchMethod
         {
-            [Fact]
-            public void Should_Throw_If_Pattern_Is_Empty()
+            public sealed class WithPatternAndPredicate
             {
-                // Given
-                var fixture = new GlobberFixture();
-                var globber = fixture.CreateGlobber();
+                [Fact]
+                public void Should_Return_Single_Path_Glob_Pattern_With_Predicate()
+                {
+                    // Given
+                    var fixture = new GlobberFixture();
+                    var globber = fixture.CreateGlobber();
 
-                // When
-                var result = Record.Exception(() => globber.Match(null));
+                    // When
+                    var result = globber.Match(
+                        "./**/*.txt",
+                        predicate => predicate.Path is DirectoryPath || predicate.Path.FullPath == "/Temp/Hello/World/Text.txt").ToArray();
 
-                // Then
-                Assert.IsArgumentNullException(result, "pattern");
+                    // Then
+                    Assert.Equal(1, result.Length);
+                    Assert.IsType<FilePath>(result[0]);
+                    Assert.Equal("/Temp/Hello/World/Text.txt", result[0].FullPath);
+                }
             }
 
-            [Fact]
-            public void Can_Traverse_Recursivly()
+            public sealed class WithPattern
             {
-                // Given
-                var fixture = new GlobberFixture();
-                var globber = fixture.CreateGlobber();
+                public sealed class WindowsSpecific
+                {
+                    [Fact]
+                    public void Will_Fix_Root_If_Drive_Is_Missing_By_Using_The_Drive_From_The_Working_Directory()
+                    {
+                        // Given
+                        var fixture = new GlobberFixture(windows: true);
+                        fixture.SetWorkingDirectory("C:/Working/");
+                        var globber = fixture.CreateGlobber();
 
-                // When
-                var result = globber.Match("/Temp/**/*.txt").ToArray();
+                        // When
+                        var result = globber.Match("/Temp/Hello/World/Text.txt").ToArray();
 
-                // Then
-                Assert.Equal(2, result.Length);
-                Assert.True(result.Any(p => p.FullPath == "/Temp/Hello/World/Text.txt"));
-                Assert.True(result.Any(p => p.FullPath == "/Temp/Goodbye/OtherText.txt"));
-            }
+                        // Then
+                        Assert.Equal(1, result.Length);
+                        Assert.Equal("C:/Temp/Hello/World/Text.txt", result[0].FullPath);
+                    }
 
-            [Fact]
-            public void Will_Append_Relative_Root_With_Implicit_Working_Directory()
-            {
-                // Given
-                var fixture = new GlobberFixture();
-                var globber = fixture.CreateGlobber();
+                    [Fact]
+                    public void Should_Throw_If_Unc_Root_Was_Encountered()
+                    {
+                        // Given
+                        var fixture = new GlobberFixture(windows: true);
+                        var globber = fixture.CreateGlobber();
 
-                // When
-                var result = globber.Match("Hello/World/Text.txt").ToArray();
+                        // When
+                        var result = Record.Exception(() => globber.Match("//Hello/World/Text.txt"));
 
-                // Then
-                Assert.Equal(1, result.Length);
-                Assert.Equal("/Temp/Hello/World/Text.txt", result[0].FullPath);
-            }
+                        // Then
+                        Assert.IsType<NotSupportedException>(result);
+                        Assert.Equal("UNC paths are not supported.", result.Message);
+                    }
 
-            [Fact]
-            public void Will_Fix_Root_If_Drive_Is_Missing_By_Using_The_Drive_From_The_Working_Directory()
-            {
-                // Given
-                var fixture = new GlobberFixture();
-                fixture.SetWorkingDirectory("C:/Working/");
-                var globber = fixture.CreateGlobber();
+                    [Theory]
+                    [InlineData(true, false)]
+                    [InlineData(false, true)]
+                    public void Should_Ignore_Case_Sensitivity_On_Case_Insensitive_Operative_System(
+                        bool isFileSystemCaseSensitive, bool shouldFindFile)
+                    {
+                        // Given
+                        var fixture = new GlobberFixture(windows: !isFileSystemCaseSensitive);
+                        var globber = fixture.CreateGlobber();
 
-                // When
-                var result = globber.Match("/Temp/Hello/World/Text.txt").ToArray();
+                        // When
+                        var result = globber.Match("/Temp/**/text.txt").ToArray();
 
-                // Then
-                Assert.Equal(1, result.Length);
-                Assert.Equal("C:/Temp/Hello/World/Text.txt", result[0].FullPath);
-            }
+                        // Then
+                        Assert.Equal(shouldFindFile, result.Length == 1);
+                    }
+                }
 
-            [Fact]
-            public void Should_Throw_If_Unc_Root_Was_Encountered()
-            {
-                // Given
-                var fixture = new GlobberFixture();
-                var globber = fixture.CreateGlobber();
+                [Fact]
+                public void Should_Throw_If_Pattern_Is_Empty()
+                {
+                    // Given
+                    var fixture = new GlobberFixture();
+                    var globber = fixture.CreateGlobber();
 
-                // When
-                var result = Record.Exception(() => globber.Match("//Hello/World/Text.txt"));
+                    // When
+                    var result = Record.Exception(() => globber.Match(null));
 
-                // Then
-                Assert.IsType<NotSupportedException>(result);
-                Assert.Equal("UNC paths are not supported.", result.Message);
-            }
+                    // Then
+                    Assert.IsArgumentNullException(result, "pattern");
+                }
 
-            [Theory]
-            [InlineData(true, false)]
-            [InlineData(false, true)]
-            public void Should_Ignore_Case_Sensitivity_On_Case_Insensitive_Operative_System(bool isFileSystemCaseSensitive, bool shouldFindFile)
-            {
-                // Given
-                var fixture = new GlobberFixture(isFileSystemCaseSensitive);
-                var globber = fixture.CreateGlobber();
+                [Fact]
+                public void Can_Traverse_Recursivly()
+                {
+                    // Given
+                    var fixture = new GlobberFixture();
+                    var globber = fixture.CreateGlobber();
 
-                // When
-                var result = globber.Match("/Temp/**/text.txt").ToArray();
+                    // When
+                    var result = globber.Match("/Temp/**/*.txt").ToArray();
 
-                // Then
-                Assert.Equal(shouldFindFile, result.Length == 1);
-            }
+                    // Then
+                    Assert.Equal(2, result.Length);
+                    Assert.True(result.Any(p => p.FullPath == "/Temp/Hello/World/Text.txt"));
+                    Assert.True(result.Any(p => p.FullPath == "/Temp/Goodbye/OtherText.txt"));
+                }
 
-            [Fact]
-            public void Should_Return_Single_Path_For_Absolute_File_Path_Without_Glob_Pattern()
-            {
-                // Given
-                var fixture = new GlobberFixture();
-                var globber = fixture.CreateGlobber();
+                [Fact]
+                public void Will_Append_Relative_Root_With_Implicit_Working_Directory()
+                {
+                    // Given
+                    var fixture = new GlobberFixture();
+                    var globber = fixture.CreateGlobber();
 
-                // When
-                var result = globber.Match("/Temp/Hello/World/Text.txt").ToArray();
+                    // When
+                    var result = globber.Match("Hello/World/Text.txt").ToArray();
 
-                // Then
-                Assert.Equal(1, result.Length);
-                Assert.IsType<FilePath>(result[0]);
-                Assert.Equal("/Temp/Hello/World/Text.txt", result[0].FullPath);
-            }
+                    // Then
+                    Assert.Equal(1, result.Length);
+                    Assert.Equal("/Temp/Hello/World/Text.txt", result[0].FullPath);
+                }
 
-            [Fact]
-            public void Should_Return_Single_Path_For_Absolute_Directory_Path_Without_Glob_Pattern()
-            {
-                // Given
-                var fixture = new GlobberFixture();
-                var globber = fixture.CreateGlobber();
+                [Fact]
+                public void Should_Return_Single_Path_For_Absolute_File_Path_Without_Glob_Pattern()
+                {
+                    // Given
+                    var fixture = new GlobberFixture();
+                    var globber = fixture.CreateGlobber();
 
-                // When
-                var result = globber.Match("/Temp/Hello/World").ToArray();
+                    // When
+                    var result = globber.Match("/Temp/Hello/World/Text.txt").ToArray();
 
-                // Then
-                Assert.Equal(1, result.Length);
-                Assert.IsType<DirectoryPath>(result[0]);
-                Assert.Equal("/Temp/Hello/World", result[0].FullPath);
-            }
+                    // Then
+                    Assert.Equal(1, result.Length);
+                    Assert.IsType<FilePath>(result[0]);
+                    Assert.Equal("/Temp/Hello/World/Text.txt", result[0].FullPath);
+                }
 
-            [Fact]
-            public void Should_Return_Single_Path_For_Relative_File_Path_Without_Glob_Pattern()
-            {
-                // Given
-                var fixture = new GlobberFixture();
-                fixture.SetWorkingDirectory("/Temp/Hello");
-                var globber = fixture.CreateGlobber();
+                [Fact]
+                public void Should_Return_Single_Path_For_Absolute_Directory_Path_Without_Glob_Pattern()
+                {
+                    // Given
+                    var fixture = new GlobberFixture();
+                    var globber = fixture.CreateGlobber();
 
-                // When
-                var result = globber.Match("./World/Text.txt").ToArray();
+                    // When
+                    var result = globber.Match("/Temp/Hello/World").ToArray();
 
-                // Then
-                Assert.Equal(1, result.Length);
-                Assert.IsType<FilePath>(result[0]);
-                Assert.Equal("/Temp/Hello/World/Text.txt", result[0].FullPath);
-            }
+                    // Then
+                    Assert.Equal(1, result.Length);
+                    Assert.IsType<DirectoryPath>(result[0]);
+                    Assert.Equal("/Temp/Hello/World", result[0].FullPath);
+                }
 
-            [Fact]
-            public void Should_Return_Single_Path_For_Relative_Directory_Path_Without_Glob_Pattern()
-            {
-                // Given
-                var fixture = new GlobberFixture();
-                fixture.SetWorkingDirectory("/Temp/Hello");
-                var globber = fixture.CreateGlobber();
+                [Fact]
+                public void Should_Return_Single_Path_For_Relative_File_Path_Without_Glob_Pattern()
+                {
+                    // Given
+                    var fixture = new GlobberFixture();
+                    fixture.SetWorkingDirectory("/Temp/Hello");
+                    var globber = fixture.CreateGlobber();
 
-                // When
-                var result = globber.Match("./World").ToArray();
+                    // When
+                    var result = globber.Match("./World/Text.txt").ToArray();
 
-                // Then
-                Assert.Equal(1, result.Length);
-                Assert.IsType<DirectoryPath>(result[0]);
-                Assert.Equal("/Temp/Hello/World", result[0].FullPath);
-            }
+                    // Then
+                    Assert.Equal(1, result.Length);
+                    Assert.IsType<FilePath>(result[0]);
+                    Assert.Equal("/Temp/Hello/World/Text.txt", result[0].FullPath);
+                }
 
-            [Fact]
-            public void Should_Return_Single_Path_Glob_Pattern_With_Predicate()
-            {
-                // Given
-                var fixture = new GlobberFixture();
-                var globber = fixture.CreateGlobber();
+                [Fact]
+                public void Should_Return_Single_Path_For_Relative_Directory_Path_Without_Glob_Pattern()
+                {
+                    // Given
+                    var fixture = new GlobberFixture();
+                    fixture.SetWorkingDirectory("/Temp/Hello");
+                    var globber = fixture.CreateGlobber();
 
-                // When
-                var result = globber.Match(
-                    "./**/*.txt",
-                    predicate=>predicate.Path is DirectoryPath || predicate.Path.FullPath == "/Temp/Hello/World/Text.txt").ToArray();
+                    // When
+                    var result = globber.Match("./World").ToArray();
 
-                // Then
-                Assert.Equal(1, result.Length);
-                Assert.IsType<FilePath>(result[0]);
-                Assert.Equal("/Temp/Hello/World/Text.txt", result[0].FullPath);
+                    // Then
+                    Assert.Equal(1, result.Length);
+                    Assert.IsType<DirectoryPath>(result[0]);
+                    Assert.Equal("/Temp/Hello/World", result[0].FullPath);
+                }
+
+                [Fact]
+                public void Should_Only_Return_Folders_For_Pattern_Ending_With_Recursive_Wildcard()
+                {
+                    // Given
+                    var fixture = new GlobberFixture();
+                    var globber = fixture.CreateGlobber();
+
+                    // When
+                    var result = globber.Match("/Temp/Hello/**").ToArray();
+
+                    // Then
+                    Assert.Equal(1, result.Length);
+                    Assert.IsType<DirectoryPath>(result[0]);
+                    Assert.Equal("/Temp/Hello/World", result[0].FullPath);
+                }
+
+                [Fact]
+                public void Should_Return_Files_And_Folders_For_Pattern_Ending_With_Wildcard()
+                {
+                    // Given
+                    var fixture = new GlobberFixture();
+                    var globber = fixture.CreateGlobber();
+
+                    // When
+                    var result = globber.Match("/Temp/**/*").ToArray();
+
+                    // Then
+                    Assert.Equal(7, result.Length);
+                }
+
+                [Fact]
+                public void Should_Return_File_For_Pattern_Ending_With_Extension_Regex()
+                {
+                    // Given
+                    var fixture = new GlobberFixture();
+                    var globber = fixture.CreateGlobber();
+
+                    // When
+                    var result = globber.Match("/Temp/**/*.txt").ToArray();
+
+                    // Then
+                    Assert.Equal(2, result.Length);
+                }
             }
         }
     }