--- a/src/Cake.Core/IO/Globber.cs
+++ b/src/Cake.Core/IO/Globber.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text.RegularExpressions;
 using Cake.Core.IO.Globbing;
 
 namespace Cake.Core.IO