--- a/src/Cake.Core/IO/Globbing/GlobTokenKind.cs
+++ b/src/Cake.Core/IO/Globbing/GlobTokenKind.cs
@@ -5,7 +5,7 @@
 
 namespace Cake.Core.IO.Globbing
 {
-    internal enum TokenKind
+    internal enum GlobTokenKind
     {
         Wildcard,
         CharacterWildcard,