… src/Cake.Core/IO/Globbing/GlobNode.cs
… src/Cake.Core/IO/Globbing/GlobNodeRewriter.cs
… src/Cake.Core/IO/Globbing/GlobParser.cs
… src/Cake.Core/IO/Globbing/GlobParserContext.cs
… src/Cake.Core/IO/Globbing/GlobToken.cs
… src/Cake.Core/IO/Globbing/GlobTokenizer.cs
… src/Cake.Core/IO/Globbing/GlobVisitor.cs
… src/Cake.Core/IO/Globbing/GlobVisitorContext.cs
… src/Cake.Core/IO/Globbing/Node.cs
… src/Cake.Core/IO/Globbing/Nodes/IdentifierNode.cs
… src/Cake.Core/IO/Globbing/Nodes/MatchableNode.cs
… src/Cake.Core/IO/Globbing/Nodes/PathSegment.cs
… src/Cake.Core/IO/Globbing/Nodes/RecursiveWildcardSegment.cs
… src/Cake.Core/IO/Globbing/Nodes/RelativeRoot.cs
… src/Cake.Core/IO/Globbing/Nodes/SegmentNode.cs
… src/Cake.Core/IO/Globbing/Nodes/UnixRoot.cs
… src/Cake.Core/IO/Globbing/Nodes/WildcardNode.cs
… src/Cake.Core/IO/Globbing/Nodes/WildcardSegment.cs
… src/Cake.Core/IO/Globbing/Nodes/WildcardSegmentNode.cs
… src/Cake.Core/IO/Globbing/Nodes/WindowsRoot.cs
… src/Cake.Testing/Cake.Testing.csproj
… src/Cake.Testing/ExceptionAsserts.cs
… src/Cake.Testing/Fakes/FakeDirectory.cs
… src/Cake.Testing/Fakes/FakeEnvironment.cs
… src/Cake.Testing/PathAsserts.cs
… src/Cake.Testing/Settings.StyleCop
… src/Settings.StyleCop