@patriksvensson 2015-08-03
Merged.